pyproject.toml
[tool.poetry]
name = "fantasy-football"
version = "0.1.0"
description = ""
authors = ["DeadlyChambers85 <shanechambers85@gmail.com>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
persian_wordcloud = "^1.3.1"
nltk = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
